run the tests in a dedicated working directory
For each iteration of each test, a special directory
is created under target/workdir. Inside this directory,
there will be stdout.txt and stderr.txt, to which Swarm's
stdout/stderr will be redirected, and a tmp directory
that will be used as java.io.tmpdir of the Swarm process.

This itself removes most of the clutter from the test tool
stdout. The remaining clutter, the PID and RSS, are removed
as well, so that the test tool stdout only shows progress.