various improvements (#1)
* add IntelliJ files to .gitignore

* improve the core loop of single test iteration

The core loop in the runTest method was rewritten to make
the code more clear and also to fix two issues:

1. When the server almost finished booting and is about
   to start deploying the application, the HTTP endpoint
   has already been exposed. It will return 404 responses
   until the application is deployed. Before this commit,
   a 404 response was considered an error and the process
   was killed, which is wrong.

2. In case "destroyForcibly" is needed, it should be waited
   upon, otherwise another iteration of the test can start
   while the process is still running. Unlikely, but possible.

* use non-deprecated Option builder

* run the tests in a dedicated working directory

For each iteration of each test, a special directory
is created under target/workdir. Inside this directory,
there will be stdout.txt and stderr.txt, to which Swarm's
stdout/stderr will be redirected, and a tmp directory
that will be used as java.io.tmpdir of the Swarm process.

This itself removes most of the clutter from the test tool
stdout. The remaining clutter, the PID and RSS, are removed
as well, so that the test tool stdout only shows progress.

* add more measurements and improve existing ones

The startup time was previously measured with precision
of 100ms, which can make quite some difference. This is
now changed to 20ms.

These measurements were added: used Java heap (measured
using jstat -gc), the uberjar size, temporary directory
usage (the WildFly server started by Swarm uses that for
various things), shutdown time.

* various small code improvements

Fixed some typos, shortened some routines, no behavior change.

* rewrite CSVCollector to use actual CSV printer

Previously, CSVCollector assembled the CSV manually. Changing
it to use CSVPrinter is more clear and future-proof.

Also, the CSVs are now always read/written as UTF-8 instead of
the default encoding.