improve the ability to skip tests
When tests are skipped, the output file (if specified)
shouldn't get overwritten. For that, the output collector
will be set to the stdout variant.

For performing the result comparison, the output collector
should no longer be checked if it creates a CSV, because
when tests are skipped, the stdout collector is used,
yet a valid output file can be set.