improve how CSV column names are handled
The Measure enum now has methods to generate CSV column names,
which is better than having the same string concatenation code
on multiple places.

Also some places of the code still used integer indexes of some
CSV columns, which is easy to break when adding new columns.
One of previous commits ensured that CSVs are always parsed
with a header, so values in CSV records can be looked up
by the column name. This is more readable and more future-proof.