Small improvements (#2)
* improve the ability to skip tests

When tests are skipped, the output file (if specified)
shouldn't get overwritten. For that, the output collector
will be set to the stdout variant.

For performing the result comparison, the output collector
should no longer be checked if it creates a CSV, because
when tests are skipped, the stdout collector is used,
yet a valid output file can be set.

* parse the CSV header when loading the CSV files for comparison

The CSV parser is created with an option to parse the header.
This means that the FailFastComparator no longer has to skip
the header manually.

* add mean, standard deviation and median to the CVS results

* add the ability to generate a comparison CSV file

This CSV has a format suitable for loading into a spreadsheet
and create comparison graphs. That is: lines are alternating
between previous and current version of each analyzed example,
and the version is printed in the first column.

* measure used heap space correctly

The Jstat class that is responsible for measuring the used
Java heap space was overly pessimistic. Common tools like
jconsole, jvisualvm and jmc all report "used heap" as a sum
of young gen and old gen, while the Jstat class also added
used metaspace and ccspace. That is doubly wrong: metaspace
isn't part of heap, and ccspace is part of metaspace, so it
was counted twice.

With this commit, the Jstat class measures used heap in
exactly the same way as other tools: young gen + old gen,
i.e. eden space + survivor spaces + old space.

* perform a warmup and a GC before measuring memory

Previously, used memory (both native RSS and Java heap)
were measured right after the first request. This isn't
very accurate. With this commit, used memory is only
measured after a small warmup is finished (100 requests)
and a Java GC is triggered (using jcmd).

* improve how CSV column names are handled

The Measure enum now has methods to generate CSV column names,
which is better than having the same string concatenation code
on multiple places.

Also some places of the code still used integer indexes of some
CSV columns, which is easy to break when adding new columns.
One of previous commits ensured that CSVs are always parsed
with a header, so values in CSV records can be looked up
by the column name. This is more readable and more future-proof.