fix usage of the datasources example
The `datasources-subsystem` example no longer exists. Instead,
the `datasources-war` example is used. That example, however,
needs to be started in one of 3 profiles; the default profile
doesn't include any datasource and so the example doesn't work.

To be able to do this, this commit adds the ability to specify
arguments to the `java -jar *-swarm.jar` commands in a special
file called `swarm-app-args.properties`.